python/requirements.txt
pyserial==3.5
fastapi==0.85.0
uvicorn[standard]==0.18.3